Attempted to start on a class for the diver
Set up the bare minimum frame work for it. Need to read up more on event
listening so I can figure out the move method. From cursory research it
looks like there might need to be a specific "update" method also that
is supposed to be called after every move but I'll look into that later.