DiverFUCKINGWORK has a moving diver
Save scubadiver1.png in the same directory as the previous code. Move
the diver with the arrow keys. The diver is rather huge as you can see
but I have the diver in the game and the diver moves.

The next step is to put this all in a class I think but I couldn't get
that to work in __init__ so that might be something to attempt later on.
For the demo on Wedneday we should probably get the treasure in and some
fish swimming around. God mode would be to get the collision going for
the fish too by then I think.

KeyboardMovement.py is a template I found online that I modified
slightly to use the image instead of circles.